Refactor program utilities classes
