+++ src/modules/logos.py
@@ -1,0 +1,106 @@
+NETWATCH_LOGO = '''
+                    :!?Y5PGGPP5!:             .!JPGBGPY7^.
+                  ..J#@#Y!75#@@@&P7:     .~7?5#@@&BY77YG#BY:
+                  .^7~^.    .!P&@@@&BJ: .7B&@@@@G7.     :!Y!
+                               :?G&@&Y: :5@&#BJ^       ..
+                   !J!.   .:~!~^  :~^.   .^^:..:~!!~:  ..YJ.
+                   7B~ .~5#&@@@&BJ:  .:  .. .JB&@@@@&G~  ~#!
+                  .77. !#@@@@@@@@&7  ?Y  J? ^B@@@@&&&@B: :!:
+                     .^JY?!~~!?PB!   5J  ?P. !BP?!^::^!!^
+                    ::.         .^.~YJ:  .?Y!~:         .:.
+                 .^JG~           :5?:  ..  .75^          ?G?:
+               .:J&@Y            !Y   ....   ?~          :B@B~.
+               ^5@@@G^       .:!J5G57.   .~YGG5?!:.     .^G@@@?
+              .J&@@@@#57~~!YBB&@@@@@@BJ7JG@@@@@@@&BPJ775B#@@@&P.
+               :#@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@B7
+               .J@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@GJ:
+                .!P#&&&@@@@@@@@@&&&&@@@@@@@@&#&&@@@@@@@@@&P?^.
+                   .:^^~!!!!YB@&&G5J?7!!?JYJ?JPPPPJ777!~~^.
+                             :Y#@@@@&#BBBB#&@@&#J^
+                               :75B&&&&@@&&&BY!:.
+                                  .:::^^^^:::.
+===================================================================================
+       ███╗   ██╗███████╗████████╗██╗    ██╗ █████╗ ████████╗ ██████╗██╗  ██╗
+       ████╗  ██║██╔════╝╚══██╔══╝██║    ██║██╔══██╗╚══██╔══╝██╔════╝██║  ██║
+       ██╔██╗ ██║█████╗     ██║   ██║ █╗ ██║███████║   ██║   ██║     ███████║
+       ██║╚██╗██║██╔══╝     ██║   ██║███╗██║██╔══██║   ██║   ██║     ██╔══██║
+       ██║ ╚████║███████╗   ██║   ╚███╔███╔╝██║  ██║   ██║   ╚██████╗██║  ██║
+       ╚═╝  ╚═══╝╚══════╝   ╚═╝    ╚══╝╚══╝ ╚═╝  ╚═╝   ╚═╝    ╚═════╝╚═╝  ╚═╝
+===================================================================================
+    v{version}                                      Noah Sickels (@NCSickels)
+===================================================================================
+'''
+
+AUTOATTACK_LOGO = '''
+===================================================================================
+                    █████╗ ██╗   ██╗████████╗ ██████╗
+                   ██╔══██╗██║   ██║╚══██╔══╝██╔═══██╗
+                   ███████║██║   ██║   ██║   ██║   ██║
+                   ██╔══██║██║   ██║   ██║   ██║   ██║
+                   ██║  ██║╚██████╔╝   ██║   ╚██████╔╝
+                   ╚═╝  ╚═╝ ╚═════╝    ╚═╝    ╚═════╝
+             █████╗ ████████╗████████╗ █████╗  ██████╗██╗  ██╗
+            ██╔══██╗╚══██╔══╝╚══██╔══╝██╔══██╗██╔════╝██║ ██╔╝
+            ███████║   ██║      ██║   ███████║██║     █████╔╝
+            ██╔══██║   ██║      ██║   ██╔══██║██║     ██╔═██╗
+            ██║  ██║   ██║      ██║   ██║  ██║╚██████╗██║  ██╗
+            ╚═╝  ╚═╝   ╚═╝      ╚═╝   ╚═╝  ╚═╝ ╚═════╝╚═╝  ╚═╝
+===================================================================================
+'''
+
+INFO_LOGO = '''
+===================================================================================
+                        ██╗███╗   ██╗███████╗ ██████╗
+                        ██║████╗  ██║██╔════╝██╔═══██╗
+                        ██║██╔██╗ ██║█████╗  ██║   ██║
+                        ██║██║╚██╗██║██╔══╝  ██║   ██║
+                        ██║██║ ╚████║██║     ╚██████╔╝
+                        ╚═╝╚═╝  ╚═══╝╚═╝      ╚═════╝
+===================================================================================
+'''
+
+NMAP_LOGO = '''
+===================================================================================
+                        ███╗   ██╗███╗   ███╗ █████╗ ██████╗
+                        ████╗  ██║████╗ ████║██╔══██╗██╔══██╗
+                        ██╔██╗ ██║██╔████╔██║███████║██████╔╝
+                        ██║╚██╗██║██║╚██╔╝██║██╔══██║██╔═══╝
+                        ██║ ╚████║██║ ╚═╝ ██║██║  ██║██║
+                        ╚═╝  ╚═══╝╚═╝     ╚═╝╚═╝  ╚═╝╚═╝
+===================================================================================
+'''
+
+PYDISCOVER_LOGO = '''
+██████╗ ██╗   ██╗██████╗ ██╗███████╗ ██████╗ ██████╗ ██╗   ██╗███████╗██████╗
+██╔══██╗╚██╗ ██╔╝██╔══██╗██║██╔════╝██╔════╝██╔═══██╗██║   ██║██╔════╝██╔══██╗
+██████╔╝ ╚████╔╝ ██║  ██║██║███████╗██║     ██║   ██║██║   ██║█████╗  ██████╔╝
+██╔═══╝   ╚██╔╝  ██║  ██║██║╚════██║██║     ██║   ██║╚██╗ ██╔╝██╔══╝  ██╔══██╗
+██║        ██║   ██████╔╝██║███████║╚██████╗╚██████╔╝ ╚████╔╝ ███████╗██║  ██║
+╚═╝        ╚═╝   ╚═════╝ ╚═╝╚══════╝ ╚═════╝ ╚═════╝   ╚═══╝  ╚══════╝╚═╝  ╚═╝
+'''
+
+SAGEMODE_LOGO_TEXT = '''
+███████╗ █████╗  ██████╗ ███████╗███╗   ███╗ ██████╗ ██████╗ ███████╗
+██╔════╝██╔══██╗██╔════╝ ██╔════╝████╗ ████║██╔═══██╗██╔══██╗██╔════╝
+███████╗███████║██║  ███╗█████╗  ██╔████╔██║██║   ██║██║  ██║█████╗
+╚════██║██╔══██║██║   ██║██╔══╝  ██║╚██╔╝██║██║   ██║██║  ██║██╔══╝
+███████║██║  ██║╚██████╔╝███████╗██║ ╚═╝ ██║╚██████╔╝██████╔╝███████╗
+╚══════╝╚═╝  ╚═╝ ╚═════╝ ╚══════╝╚═╝     ╚═╝ ╚═════╝ ╚═════╝ ╚══════╝
+'''
+
+SAGEMODE_LOGO = '''
+                         @@@%%%%%@@@
+                     @%##`````@`````##&@
+                   @##````````@````````##@
+                 @%#`````````@@@`````````#%@
+                 &#``````````@@@``````````#&
+                @#````@@@@@@@@@@@@@@@@@````#@
+                @%@``@@@@@@@@@@@@@@@@@@@``@%@
+                @%@```@@@@@@@@@@@@@@@@@```#%@
+                @@# `````````@@@``````````#@@
+                 &#``````````@@@``````````#&
+                  @##`````````@`````````##@
+                    @##```````@``````###@
+                       @@#````@````#@@
+                         @@@%%%%%@@@
+'''
